added pvwatts model to detect anomalies in outlier script
