+++ pvanalytics/quality/outliers.py
@@ -2,6 +2,8 @@
 import pandas as pd
 from scipy import stats
 from statsmodels import robust
+import pvlib
+from pvlib.temperature import TEMPERATURE_MODEL_PARAMETERS
 
 
 def tukey(data, k=1.5):
@@ -124,3 +126,277 @@ def hampel(data, window=5, max_deviation=3.0, scale=None):
         kwargs=kwargs
     )
     return deviation > max_deviation * mad
+
+
+def run_pvwatts_data_checks(power_series, nsrdb_weather_df):
+    """Check that the passed parameters can be run through the function.
+    
+    This includes checking the passed time series to ensure it has a
+    datetime index, and that its frequency matches with nsrdb weather
+    data frequency.
+
+    Parameters
+    ----------
+    power_series : Pandas series with datetime index.
+        Time series of a PV power data stream.
+    nsrdb_weather_df : Pandas dataframe with datetime index
+        A pandas dataframe containing NSRDB PSM3 weather 'Temperature', 'DHI',
+        'DNI', and 'Wind Speed' columns with datetime index. The minimum
+        and maximum datetime matches the minimum and maximum datetime of the
+        actual data.
+
+    Returns
+    -------
+    daily_series : Pandas series with datetime index.
+        Daily time series of a PV power data stream. This series represents
+        the summed daily values of the particular data stream.
+    """
+    # Check that the time series has a datetime index, and consists of numeric
+    # values
+    if not isinstance(power_series.index, pd.DatetimeIndex):
+        raise TypeError('Must be a Pandas series with a datetime index.')
+    # Get the frequency of weather data
+    nsrdb_weather_df = pd.infer_freq(nsrdb_weather_df.index)
+    # Resample power series to same frequency as weather data
+    power_series = power_series.resample(nsrdb_weather_df).mean()
+    return power_series
+
+
+def run_pvwatts_model(tilt, azimuth, dc_capacity, dc_inverter_limit,
+                      solar_zenith, solar_azimuth, dni, dhi, ghi, dni_extra,
+                      relative_airmass, temperature, wind_speed,
+                      temperature_model_parameters,
+                      temperature_coefficient, tracking):
+    r"""Run the PVWatts model with NSRDB data across the time period as inputs.
+
+    Parameters
+    ----------
+    tilt : Float
+        Tilt angle of site in degrees.
+    azimuth : Float
+        Azimuth angle of site in degrees.
+    dc_capacity : Float
+        DC capacity of the inverter in kW.
+    dc_inverter_limit : Float
+        Maximum DC capacity for the inverter in kW.
+    solar_zenith : Series
+        Solar zenith angle in degrees
+    solar_azimuth :
+        Azimuth angle of the sun in degrees East of North
+    dni : Series
+        Direct normal irradiance in :math:`W/m^2`
+    dhi : Series
+        Diffuse horizontal irradiance in :math:`W/m^2`
+    ghi : Series
+        Global horizontal irradiance in :math:`W/m^2`
+    dni_extra : Series
+        Extraterrestrial normal irradiance in :math:`W/m^2`
+    relative_airmass : Series
+        Relative airmass (not adjusted for pressure)
+    temperature : Series
+        Ambient dry bulb temperature in C. 
+    wind_speed : Series
+        Wind speed at a height of 10 meters in :math:`m/s`.
+    temperature_model_parameters : Dict
+        Parameters for the cell temperature model.
+    temperature_coefficient : Float
+        Temperature coefficient of DC power. [1/C]
+    tracking : Boolean
+        True if tracking. Otherwise, False if not tracking.
+
+    Returns
+    -------
+    pdc : Series
+        Pandas series of DC power output in kW modeled from PVWatts with
+        datetime index.
+    """
+    if tracking:
+        tracker_angles = pvlib.tracking.singleaxis(solar_zenith,
+                                                   solar_azimuth,
+                                                   axis_tilt=tilt,
+                                                   axis_azimuth=azimuth,
+                                                   backtrack=True,
+                                                   gcr=0.4, max_angle=60)
+        surface_tilt = tracker_angles['surface_tilt']
+        surface_azimuth = tracker_angles['surface_azimuth']
+    else:
+        surface_tilt = tilt
+        surface_azimuth = azimuth
+
+    poa = pvlib.irradiance.get_total_irradiance(
+        surface_tilt, surface_azimuth,
+        solar_zenith,
+        solar_azimuth,
+        dni, ghi, dhi,
+        dni_extra=dni_extra,
+        airmass=relative_airmass,
+        albedo=0.2,
+        model='perez'
+    )
+
+    aoi = pvlib.irradiance.aoi(surface_tilt, surface_azimuth,
+                               solar_zenith, solar_azimuth)
+    # Run IAM model
+    iam = pvlib.iam.physical(aoi, n=1.5)
+    # Apply IAM to direct POA component only
+    poa_transmitted = poa['poa_direct'] * iam + poa['poa_diffuse']
+    temp_cell = pvlib.temperature.sapm_cell(
+        poa['poa_global'],
+        temperature,
+        wind_speed,
+        **temperature_model_parameters
+    )
+    pdc = pvlib.pvsystem.pvwatts_dc(
+        poa_transmitted,
+        temp_cell,
+        dc_capacity,
+        temperature_coefficient
+    )
+    return pdc
+
+
+def build_pvwatts_model(lat, long, tilt, azimuth, power, tracking,
+                        nsrdb_weather_df):
+    """Models the system's power output using the PVWatts model.
+    
+    Pulls the NSRDB weather data via NSRDB PSM3 api and uses the system's
+    metadata to build the PVWatts model to get the predicted power ourtput.
+    
+    Parameters
+    ----------
+    lat : Float
+        Latitude value of site.
+    long : FLoat
+        Longitude value of site.
+    tilt : Float
+        Tilt angle of site in degrees.
+    azimuth : Float
+        Azimuth angle of site in degrees.
+    power : Float
+        DC capacity of the inverter in kW.
+    tracking : Boolean
+        True if tracking. Otherwise, False if not tracking.
+    nsrdb_weather_df : Pandas dataframe with datetime index
+        A pandas dataframe containing NSRDB PSM3 weather 'Temperature', 'DHI',
+        'DNI', and 'Wind Speed' columns with datetime index. The minimum
+        and maximum datetime matches the minimum and maximum datetime of the
+        actual data.
+
+    Returns
+    -------
+    master_df : Pandas dataframe with datetime index.
+        A pandas dataframe containing columns 'actual_power_kW',
+        'predicted_power_kW', 'percent_difference', and 'anomalous'
+        columns and datetime index.
+    """
+
+    # Build out the PVWatts model
+    solpos = pvlib.solarposition.get_solarposition(
+        nsrdb_weather_df.index, lat, long)
+    dni_extra = pvlib.irradiance.get_extra_radiation(
+        nsrdb_weather_df.index)
+    relative_airmass = pvlib.atmosphere.get_relative_airmass(solpos.zenith)
+    temp_params = TEMPERATURE_MODEL_PARAMETERS['sapm']['open_rack_glass_glass']
+    pdc = run_pvwatts_model(tilt=tilt,
+                            azimuth=azimuth,
+                            dc_capacity=power,
+                            dc_inverter_limit=power * 1.5,
+                            solar_zenith=solpos.zenith,
+                            solar_azimuth=solpos.azimuth,
+                            dni=nsrdb_weather_df['DNI'],
+                            dhi=nsrdb_weather_df['DHI'],
+                            ghi=nsrdb_weather_df['GHI'],
+                            dni_extra=dni_extra,
+                            relative_airmass=relative_airmass,
+                            temperature=nsrdb_weather_df['Temperature'],
+                            wind_speed=nsrdb_weather_df['Wind Speed'],
+                            temperature_model_parameters=temp_params,
+                            temperature_coefficient=-0.0047,
+                            tracking=tracking)
+    return pdc
+
+
+def get_anomalous_days(power_time_series, lat, long, tilt, azimuth, tracking,
+                       nsrdb_weather_df, pct_threshold=50, dc_capacity=None):
+    """Identifies anomalous days for inverter power time series.
+    
+    Compares the actual daily inverter power time series data with PVWatts model
+    power time series output by calculating their percent deference. Days where
+    the difference over the specified percent difference threshold are flagged
+    as anomalous.
+
+    Parameters
+    ----------
+    power_time_series : Pandas series with datetime index.
+        Time series of a PV power data stream. The units for power should be
+        kW.
+    lat : Float
+        Latitude value of site.
+    long : FLoat
+        Longitude value of site.
+    tilt : Float
+        Tilt angle of site in degrees.
+    azimuth : Float
+        Azimuth angle of site in degrees.
+    tracking : Boolean
+        True if tracking. Otherwise, False if not tracking.
+    nsrdb_weather_df : Pandas dataframe with datetime index
+        A pandas dataframe containing NSRDB PSM3 weather 'Temperature', 'DHI',
+        'DNI', and 'Wind Speed' columns with datetime index. The minimum
+        and maximum datetime matches the minimum and maximum datetime of the
+        actual data.
+    pct_threshold : Float
+        Percent difference threshold for flagging data as anomalies.
+        Defaulted to 50.
+    dc_capacity : None or Float
+        DC capacity of the inverter in kW. If the inverter dc capacity is not
+        known, set as None so that it can be calculated by finding the 95%
+        percentile of the passed time series.
+        Defaulted to None.
+
+    Returns
+    -------
+    master_df : Pandas dataframe with datetime index
+        A pandas dataframe containing columns 'actual_power_kW',
+        'predicted_power_kW', 'percent_difference', and 'anomalous'
+        columns with datetime index.
+
+    """
+    # Run data check on time series data to make sure index is datetime
+    # and it matches nsrdb weather data frequency
+    actual_power_ts = run_pvwatts_data_checks(power_time_series,
+                                              nsrdb_weather_df)
+    # If dc_capacity is None, calculate the 95% percentile
+    if dc_capacity is None:
+        dc_capacity = actual_power_ts.quantile(0.95)
+    # Get predicted power time series data from PVWatts
+    predicted_power_ts = build_pvwatts_model(
+        lat=lat,
+        long=long,
+        tilt=tilt,
+        azimuth=azimuth,
+        power=dc_capacity,
+        tracking=tracking,
+        nsrdb_weather_df=nsrdb_weather_df)
+    # Make timezone unaware
+    predicted_power_ts = predicted_power_ts.tz_localize(None)
+    # Combine actual and predicted power data
+    master_df = pd.DataFrame(
+        index=actual_power_ts.index,
+        data={'actual_power_kW': actual_power_ts,
+              'predicted_power_kW': predicted_power_ts})
+    master_df.index.name = "utc_measured_on"
+    # Resample to daily frequency
+    master_df = master_df.resample("D").mean() 
+    # Get percent difference
+    abs_diff = abs(master_df['actual_power_kW'] - 
+                   master_df['predicted_power_kW'])
+    mean_val = (master_df['actual_power_kW'] +
+                master_df['predicted_power_kW']) / 2
+    master_df['percent_difference'] = (abs_diff / mean_val) * 100
+    # Mark as anomalous (True) if percent difference passes the threshold
+    # Otherwise, False
+    master_df['anomalous'] = master_df['percent_difference'] > pct_threshold
+
+    return master_df
+
